+++ CHANGELOG.md
@@ -30,9 +30,15 @@ You can check your current version with the following command:
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
 
+## 2.1.0a2
+
+**Sep 2, 2024**
+- Fix test coverage for `Order::estimate`, `Order::place`, and `Order::track_status` methods.
+- Change `Order::track_status` report_time input type to float.
+
 ## 2.1.0a1
 
-**TBD, 2024**
+**Aug 30, 2024**
 - Improve test coverage `Order` class.
 
 ## 2.0.1